[Refactor] Vector3 float 타입 변환 및 Field 구조 개선
- Protobuf Struct.proto 내부 Vector3 및 yaw 타입을 double에서 float으로 변경하여 네트워크 대역폭 절약 및 클라이언트 동기화 일치.

- 서버 내부 로직(Player.cpp, Component.h 등)에서 double 연산을 float으로 수정하여 성능 최적화.

- 기존 Field.cpp/h 파일에 혼재되어 있던 FieldManager를 별도의 클래스(FieldManager.cpp/h)로 분리하고, Contents 하위에 Field 폴더를 신설하여 응집도 향상.

- nlohmann/json 라이브러리를 활용한 FieldData 파싱 및 로드 로직 구현.

- 불필요한 빌드 파일 및 설정 파일 제거.

+++ Server/GameServer/Contents/Component.h
@@ -27,11 +27,11 @@ public:
 	TransformComponent(CreateKey key, std::weak_ptr<GameObject> owner) : Component(key, owner) {}
 
 	const Vector3& GetPos() const { return _position; }
-	double GetYaw() const { return _yaw; }
+	float GetYaw() const { return _yaw; }
 
 	void SetPos(const Vector3& netPos) {_position = netPos; };
 
 private:
 	Vector3 _position;
-	double _yaw = 0;
+	float _yaw = 0;
 };
